PERA-2643 Remove standalone staking fragment
